+++ test/harness/$FAIL.js
@@ -1,0 +1,3 @@
+function $FAIL(message) {
+    testFailed(message);
+}
