+++ test/harness/fnGlobalObject.js
@@ -1,0 +1,5 @@
+//-----------------------------------------------------------------------------
+var __globalObject = Function("return this;")();
+function fnGlobalObject() {
+     return __globalObject;
+}
