+++ test/suite/ch08/8.7/8.7.2/8.7.2-3-a-2gs.js
@@ -9,10 +9,11 @@ es5id: 8.7.2-3-a-2gs
 description: >
     Strict Mode - 'runtime' error is thrown before LeftHandSide
     evaluates to an unresolvable Reference
-negative: NotEarlyError
+negative: Test262Error
 flags: [onlyStrict]
+includes: [Test262Error.js]
 ---*/
 
 "use strict";
-throw NotEarlyError;
+throw new Test262Error();
 b = 11;
