+++ test/suite/ch11/11.13/11.13.1/11.13.1-4-28gs.js
@@ -9,10 +9,11 @@ es5id: 11.13.1-4-28gs
 description: >
     Strict Mode - SyntaxError is thrown if the identifier 'Math.PI'
     appears as the LeftHandSideExpression of simple assignment(=)
-negative: NotEarlyError
+negative: Test262Error
 flags: [onlyStrict]
+includes: [Test262Error.js]
 ---*/
 
 "use strict";
-throw NotEarlyError;
+throw new Test262Error();
 Math.PI = 20;
