+++ test/suite/ch13/13.2/13.2-19-b-3gs.js
@@ -9,11 +9,12 @@ es5id: 13.2-19-b-3gs
 description: >
     StrictMode - error is thrown when assign a value to the 'caller'
     property of a function object
-negative: NotEarlyError
+negative: Test262Error
 flags: [onlyStrict]
+includes: [Test262Error.js]
 ---*/
 
 "use strict";
-throw NotEarlyError;
+throw new Test262Error();
 function _13_2_19_b_3_gs() {}
 _13_2_19_b_3_gs.caller = 1;
