+++ test/suite/ch15/15.3/15.3.2/15.3.2.1/15.3.2.1-10-6gs.js
@@ -10,9 +10,10 @@ description: >
     Strict Mode - SyntaxError is thrown if a function using the
     Function constructor has two identical parameters in (local)
     strict mode
-negative: NotEarlyError
+negative: Test262Error
 flags: [onlyStrict]
+includes: [Test262Error.js]
 ---*/
 
-throw NotEarlyError;
+throw new Test262Error();
 var _15_3_2_1_10_6_fun = new Function('param_1', 'param_2', 'param_1', '"use strict";return 0;');
