+++ tools/packaging/parseTestRecord.py
@@ -17,6 +17,8 @@ import sys
 import tempfile
 import time
 
+import yaml
+
 # from TestCasePackagerConfig import *
 
 headerPatternStr = r"(?:(?:\s*\/\/.*)?\s*\n)*"
@@ -34,6 +36,9 @@ testRecordPattern = re.compile(r"^(" + headerPatternStr +
 stars = re.compile(r"\s*\n\s*\*\s?")
 atattrs = re.compile(r"\s*\n\s*\*\s*@")
 
+yamlPattern = re.compile(r"---((?:\s|\S)*)---")
+newlinePattern = re.compile(r"\n")
+
 def stripStars(text):
     return stars.sub('\n', text).strip()
 
@@ -41,25 +46,63 @@ def stripHeader(src):
     header = headerPattern.match(src).group(0)
     return src[len(header):]
 
-def parseTestRecord(src, name):
-    testRecord = {}
+def matchParts(src, name):
     match = testRecordPattern.match(src)
     if match == None:
         raise Exception('unrecognized: ' + name)
+    return match
+
+def hasYAML(text):
+    match = yamlPattern.match(text)
+    if match == None:
+        return False
+    return True
+
+def oldAttrParser(testRecord, body, name):
+    propTexts = atattrs.split(body)
+    testRecord['commentary'] = stripStars(propTexts[0])
+    del propTexts[0]
+    for propText in propTexts:
+        propMatch = re.match(r"^\w+", propText)
+        if propMatch == None:
+            raise Exception('Malformed "@" attribute: ' + name)
+        propName = propMatch.group(0)
+        propVal = stripStars(propText[len(propName):])
+            
+        if propName in testRecord:
+            raise Exception('duplicate: ' + propName)
+        testRecord[propName] = propVal;
+
+def yamlAttrParser(testRecord, attrs, name):
+    match = yamlPattern.match(attrs)
+    body = match.group(1)
+    parsed = yaml.load(body)
+
+    if (parsed is None):
+        print "Failed to parse yaml in name %s"%(name)
+        return
+
+    for key in parsed:
+        value = parsed[key]
+        if key == "info":
+            key = "commentary"
+        testRecord[key] = value
+        
+    if 'flags' in testRecord:
+        for flag in testRecord['flags']:
+            testRecord[flag] = ""
+
+def parseTestRecord(src, name):
+    testRecord = {}
+    match = matchParts(src, name)
     testRecord['header'] = match.group(1).strip()
     testRecord['test'] = match.group(3) # do not trim
-    if match.group(2):
-        propTexts = atattrs.split(match.group(2))
-        testRecord['commentary'] = stripStars(propTexts[0])
-        del propTexts[0]
-        for propText in propTexts:
-            propMatch = re.match(r"^\w+", propText)
-            if propMatch == None:
-                raise Exception('Malformed "@" attribute: ' + name)
-            propName = propMatch.group(0)
-            propVal = stripStars(propText[len(propName):])
-            
-            if propName in testRecord:
-                raise Exception('duplicate: ' + propName)
-            testRecord[propName] = propVal;
+
+    attrs = match.group(2)
+    if attrs:
+        if hasYAML(attrs):
+            yamlAttrParser(testRecord, attrs, name)
+        else:
+            oldAttrParser(testRecord, attrs, name)
+
     return testRecord
