+++ website/harness/$PRINT.js
@@ -1,0 +1,4 @@
+//adaptors for Test262 framework
+function $PRINT(message) {
+
+}
