+++ website/harness/fnExists.js
@@ -1,0 +1,7 @@
+//-----------------------------------------------------------------------------
+function fnExists(/*arguments*/) {
+    for (var i = 0; i < arguments.length; i++) {
+        if (typeof (arguments[i]) !== "function") return false;
+    }
+    return true;
+}
