+++ website/harness/runTestCase.js
@@ -1,0 +1,5 @@
+function runTestCase(testcase) {
+    if (testcase() !== true) {
+        $ERROR("Test case returned non-true value!");
+    }
+}
